Added Spring security

+++ src/main/java/com/online/students/service/API/Assistances/AssistanceServiceImpl.java
@@ -1,10 +1,14 @@
 package com.online.students.service.API.Assistances;
 
+import com.online.students.service.API.Users.Roles;
 import com.online.students.service.API.Users.UserEntity;
 import com.online.students.service.API.Users.UserService;
 import com.online.students.service.API.AssistanceCategories.AssistanceCategoryEntity;
 import com.online.students.service.API.AssistanceCategories.AssistanceCategoryService;
+import org.springframework.http.HttpStatus;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 
@@ -34,7 +38,7 @@ public class AssistanceServiceImpl implements AssistanceService{
     @Override
     public AssistanceEntity create(AssistanceDTO assistanceDTO) {
         AssistanceCategoryEntity assistanceCategory = assistanceCategoryService.getById(assistanceDTO.idAssistanceCategory());
-        UserEntity instructor = userService.getOneById(assistanceDTO.idInstructor());
+        UserEntity instructor = userService.getOneByEmail(SecurityContextHolder.getContext().getAuthentication().getName());
 
         AssistanceEntity assistance = new AssistanceEntity(
                 assistanceDTO.title(),
@@ -49,7 +53,11 @@ public class AssistanceServiceImpl implements AssistanceService{
     @Override
     public boolean delete(Long id) {
         AssistanceEntity assistance = assistanceRepository.getReferenceById(id);
-        assistanceRepository.delete(assistance);
-        return true;
+        UserEntity currentUser = userService.getOneByEmail(SecurityContextHolder.getContext().getAuthentication().getName());
+        if (assistance.getInstructor().equals(currentUser) || currentUser.getRole().equals(Roles.ADMIN) || currentUser.getRole().equals(Roles.OWNER)) {
+            assistanceRepository.delete(assistance);
+            return true;
+        }
+        throw new ResponseStatusException(HttpStatus.FORBIDDEN, "You don't have permission to delete this assistance");
     }
 }
